requirements.txt
# requirements.txt
pyautogui
pytesseract
openai
keyboard
PillOW
python-dotenv
numpy